Update lessytracker.py
The Lessy Tracker was not working because of new Riot Id System. It is working now but there is a problem about live game datas because of Riot Game Api. According to what is written in the document, they will solve it in a short time. If you have not changed your username when the Riot Id system arrives, the Lessy Tracker will work.